+++ gateway/src/main/java/ru/practirum/shareit/exception/ErrorHandler.java
@@ -1,0 +1,44 @@
+package ru.practirum.shareit.exception;
+
+import jakarta.validation.ConstraintViolationException;
+import jakarta.validation.ValidationException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+@Slf4j
+public class ErrorHandler {
+
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ErrorResponse handleValidateException(ValidationException e) {
+        log.warn("Ошибка при валидации " + e.getMessage());
+        return new ErrorResponse("Ошибка валидации,проверьте корректность данных");
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ErrorResponse handleInvalidValidate(MethodArgumentNotValidException e) {
+        log.warn("Ошибка валидации данных " + e.getMessage());
+        return new ErrorResponse("Ошибка валидации данных");
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ErrorResponse handleInvalidValidate(ConstraintViolationException e) {
+        log.warn("Ошибка валидации данных " + e.getMessage());
+        return new ErrorResponse("Ошибка валидации данных");
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    public ErrorResponse handleInvalidValidate(Exception e) {
+        log.warn("Ошибка сервера " + e.getMessage());
+        return new ErrorResponse("Ошибка сервера");
+    }
+}
