+++ gateway/src/main/java/ru/practirum/shareit/exception/ErrorResponse.java
@@ -1,0 +1,16 @@
+package ru.practirum.shareit.exception;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ErrorResponse {
+
+    private final String error;
+
+    public ErrorResponse(String error) {
+        this.error = error;
+
+    }
+}
