+++ server/src/main/java/ru/practicum/shareit/exception/ErrorHandler.java
@@ -1,9 +1,7 @@
 package ru.practicum.shareit.exception;
 
-import jakarta.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -31,20 +29,6 @@ public class ErrorHandler {
     public ErrorResponse handleValidateException(ValidationException e) {
         log.warn("Ошибка при валидации " + e.getMessage());
         return new ErrorResponse("Ошибка валидации,проверьте корректность данных");
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleInvalidValidate(MethodArgumentNotValidException e) {
-        log.warn("Ошибка валидации данных " + e.getMessage());
-        return new ErrorResponse("Ошибка валидации данных");
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleInvalidValidate(ConstraintViolationException e) {
-        log.warn("Ошибка валидации данных " + e.getMessage());
-        return new ErrorResponse("Ошибка валидации данных");
     }
 
     @ExceptionHandler
