+++ server/src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -1,8 +1,5 @@
 package ru.practicum.shareit.item.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import ru.practicum.shareit.booking.dto.BookingDto;
 import ru.practicum.shareit.item.comment.CommentDto;
@@ -17,18 +14,13 @@ import java.util.List;
 public class ItemDto {
 
     private Long id;
-    @NotNull(message = "имя не может быть пустым")
-    @NotBlank
     private String name;
-    @NotNull(message = "описание не может быть пустым")
     private String description;
-    @NotNull
     private Boolean available;
     private Long ownerId;
     private List<CommentDto> comments;
     private BookingDto lastBooking;
     private BookingDto nextBooking;
-    @JsonProperty("requestId")
     private Long requestId;
 
 }
