+++ server/src/main/java/ru/practicum/shareit/user/dto/NewUserRequest.java
@@ -1,8 +1,6 @@
 package ru.practicum.shareit.user.dto;
 
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -14,10 +12,6 @@ import lombok.NoArgsConstructor;
 @Data
 public class NewUserRequest {
 
-    @NotNull(message = "имя не может быть пустым")
-    @NotBlank
     private String name;
-    @Email
-    @NotNull(message = "email не может быть пустым")
     private String email;
 }
