+++ server/src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -1,7 +1,5 @@
 package ru.practicum.shareit.user.dto;
 
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -13,11 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class UserDto {
 
-    @NotNull(message = "id не может быть пустым")
     private Long id;
-    @NotNull(message = "имя не может быть пустым")
     private String name;
-    @Email
-    @NotNull(message = "email не может быть пустым")
     private String email;
 }
